Feature/migrations

See merge request ods/newsystem!7

+++ database/migrations/2020_07_04_100000_create_user_types_table.php
@@ -19,7 +19,7 @@ class CreateUserTypesTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('slug');
-            $table->integer('group_id')->unsigned();
+            $table->bigInteger('group_id')->unsigned();
             $table->timestamps();
             $table->softDeletes();
